# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Auction System",
        "description": "Auction System API Documentation",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/accounts/register": {
            "post": {
                "tags": [
                    "Accounts",
                    "Accounts"
                ],
                "summary": "Create Account",
                "description": "DEPRECATED: Use /auth/register instead\n\nCreate new account (UC06)\nThis endpoint is deprecated. Use /auth/register for OTP-based registration.",
                "operationId": "create_account_api_v1_accounts_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "deprecated": true
            }
        },
        "/api/v1/accounts/profile": {
            "get": {
                "tags": [
                    "Accounts",
                    "Accounts"
                ],
                "summary": "Get User Profile",
                "description": "Get current user profile information\n\nGET /accounts/profile\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: User profile information",
                "operationId": "get_user_profile_api_v1_accounts_profile_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Accounts",
                    "Accounts"
                ],
                "summary": "Update User Profile",
                "description": "Update user profile information (UC007)\n\nPUT /accounts/profile\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nBody: { \"firstName\": \"John\", \"lastName\": \"Doe\", ... }\nReturns: Updated user profile",
                "operationId": "update_user_profile_api_v1_accounts_profile_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/accounts/change-password": {
            "post": {
                "tags": [
                    "Accounts",
                    "Accounts"
                ],
                "summary": "Change Password",
                "description": "Change user password\n\nPOST /accounts/change-password\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nBody: { \"old_password\": \"...\", \"new_password\": \"...\" }\nReturns: Success message",
                "operationId": "change_password_api_v1_accounts_change_password_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/accounts/me": {
            "get": {
                "tags": [
                    "Accounts",
                    "Accounts"
                ],
                "summary": "Get Me",
                "description": "DEPRECATED: Use /auth/me instead\n\nGet current user information",
                "operationId": "get_me_api_v1_accounts_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountCreate": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "firstName": {
                        "type": "string",
                        "title": "Firstname"
                    },
                    "lastName": {
                        "type": "string",
                        "title": "Lastname"
                    },
                    "phoneNumber": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phonenumber"
                    },
                    "dateOfBirth": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dateofbirth"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "email",
                    "password",
                    "firstName",
                    "lastName"
                ],
                "title": "AccountCreate"
            },
            "AccountStatus": {
                "type": "string",
                "enum": [
                    "active",
                    "inactive",
                    "suspended"
                ],
                "title": "AccountStatus"
            },
            "AccountUpdate": {
                "properties": {
                    "firstName": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Firstname"
                    },
                    "lastName": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Lastname"
                    },
                    "phoneNumber": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phonenumber"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "dateOfBirth": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dateofbirth"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    }
                },
                "type": "object",
                "title": "AccountUpdate"
            },
...
            "ChangePasswordRequest": {
                "properties": {
                    "old_password": {
                        "type": "string",
                        "title": "Old Password"
                    },
                    "new_password": {
                        "type": "string",
                        "title": "New Password"
                    }
                },
                "type": "object",
                "required": [
                    "old_password",
                    "new_password"
                ],
                "title": "ChangePasswordRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "success": {
                        "type": "boolean",
                        "title": "Success",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "MessageResponse"
            },
...
            "UserResponse": {
                "properties": {
                    "accountID": {
                        "type": "integer",
                        "title": "Accountid"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "firstName": {
                        "type": "string",
                        "title": "Firstname"
                    },
                    "lastName": {
                        "type": "string",
                        "title": "Lastname"
                    },
                    "phoneNumber": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phonenumber"
                    },
                    "dateOfBirth": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dateofbirth"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "role": {
                        "$ref": "#/components/schemas/UserRole"
                    },
                    "status": {
                        "$ref": "#/components/schemas/AccountStatus"
                    },
                    "lastLoginAt": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Lastloginat"
                    },
                    "isAuthenticated": {
                        "type": "boolean",
                        "title": "Isauthenticated"
                    }
                },
                "type": "object",
                "required": [
                    "accountID",
                    "username",
                    "email",
                    "firstName",
                    "lastName",
                    "role",
                    "status",
                    "isAuthenticated"
                ],
                "title": "UserResponse"
            },
            "UserRole": {
                "type": "string",
                "enum": [
                    "user",
                    "admin"
                ],
                "title": "UserRole"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}